% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/learningFunctions.R
\name{applylearning}
\alias{applylearning}
\title{Applies learning to BEV purchase cost}
\usage{
applylearning(
  non_fuel_costs,
  capcost4W,
  gdx,
  EDGE2teESmap,
  demand_learntmp,
  ES_demandpr,
  ES_demand
)
}
\arguments{
\item{non_fuel_costs}{total non fuel costs}

\item{capcost4W}{purchase prices of 4W on which learning is applied}

\item{gdx}{input gdx file}

\item{EDGE2teESmap}{mapping of EDGE-T/GCAM technologies to REMIND ES technologies}

\item{demand_learntmp}{the demand for vehicles from the previous iteration}

\item{ES_demandpr}{the ES demand of the previous iteration}

\item{ES_demand}{the ES demand of the current iteration}
}
\description{
Applies learning to BEV purchase cost
}
