% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/learningFunctions.R
\name{calc_num_vehicles_stations}
\alias{calc_num_vehicles_stations}
\title{Calculate number of vehicles scaling up the normalized demand with the aggregate total demand}
\usage{
calc_num_vehicles_stations(
  norm_dem,
  intensity,
  ES_demand_all,
  loadFactor,
  EDGE2teESmap,
  rep
)
}
\arguments{
\item{norm_dem}{normalized demand shares}

\item{intensity}{energy intensity of CES nodes, in million pkm/EJ}

\item{ES_demand_all}{total demand for ESs}

\item{loadFactor}{load factor of vehicles}

\item{EDGE2teESmap}{mapping of EDGE-T/GCAM technologies to REMIND ES technologies}

\item{rep}{if is it a stand alone run or the last iteration of the coupled run->set to TRUE}
}
\description{
Calculate number of vehicles scaling up the normalized demand with the aggregate total demand
}
