% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_capCosts.R
\name{calculate_capCosts}
\alias{calculate_capCosts}
\title{Evaluate the costs of REMIND technology placeholders (e.g. te_eselt_pass_sm).}
\usage{
calculate_capCosts(
  base_price,
  Fdemand_ES,
  stations,
  EDGE2CESmap,
  EDGE2teESmap,
  REMINDyears,
  scenario
)
}
\arguments{
\item{base_price}{base prices from logit calculations}

\item{Fdemand_ES}{normalized ES demand}

\item{stations}{costs of stations in $/km}

\item{EDGE2CESmap}{map from top level EDGE-T/GCAM categories to REMIND CES nodes}

\item{EDGE2teESmap}{mapping of EDGE-T/GCAM technologies to REMIND ES technologies}

\item{REMINDyears}{range of REMIND timesteps}

\item{scenario}{EDGE-T scenario name}
}
\description{
Evaluate the costs of REMIND technology placeholders (e.g. te_eselt_pass_sm).
}
