% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_vint.R
\name{calcVint}
\alias{calcVint}
\title{Calculate vintages composition, costs and intensity}
\usage{
calcVint(shares, totdem_regr, prices, mj_km_data, years)
}
\arguments{
\item{shares}{logit tree shares}

\item{totdem_regr}{total demand as obtained from the regression}

\item{prices}{costs of transport options in the logit tree}

\item{mj_km_data}{energy intensity of transport options in the logit tree}

\item{years}{range of years for the vintage calculation}
}
\description{
Calculate vintages composition, costs and intensity
}
