% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{collectScens}
\alias{collectScens}
\title{Collects csv from different realizations and collects them in csv files compatible with mrremind structure.}
\usage{
collectScens(output_folder)
}
\arguments{
\item{output_folder}{directory where the output files are saved}
}
\value{
saves csv files with all EDGE-transport scenarios collected
}
\description{
Collects csv from different realizations and collects them in csv files compatible with mrremind structure.
}
\author{
Marianna Rottoli
}
