% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{compareOutputs}
\alias{compareOutputs}
\title{Compare REMIND outputs for different runs}
\usage{
compareOutputs(runs, pdf_path)
}
\arguments{
\item{runs}{list of the form `scen`=<folder>, where `scen` is used to identify the run
and folder is the output folder of the run.}

\item{pdf_path}{path to the PDF file to be produced.}
}
\description{
Compare REMIND outputs for different runs
}
\author{
Alois Dirnaichner
}
