% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareScenarios_EDGET.R
\name{compareScenarios_EDGET}
\alias{compareScenarios_EDGET}
\title{Render CompareScenarios EDGE Transport}
\usage{
compareScenarios_EDGET(
  mifScen,
  mifHist,
  outputDir = getwd(),
  outputFile = "CompareScenarios_EDGE-Transport",
  outputFormat = "PDF"
)
}
\arguments{
\item{mifScen}{\code{character(n)}, optionally named. Paths to scenario mifs.
If the vector has names, those are used to refer to the scenarios in the
output file.}

\item{mifHist}{\code{character(1)}. Path to historical mif.}

\item{outputDir}{\code{character(1)}. The directory where the output document
and intermediary files are created.}

\item{outputFile}{\code{character(1)}. File name (without extension) of the
output document to be created.}

\item{outputFormat}{\code{character(1)}, not case-sensitive. \code{"html"},
\code{"pdf"}, or \code{"rmd"}.}
}
\value{
The value returned by \code{\link[rmarkdown:render]{render()}}.
}
\description{
A wrapper for piamPlotComparison::compareScenarios
}
