% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compScenEDGET.R
\name{compScenEDGET}
\alias{compScenEDGET}
\title{Read in all information from previous EDGE runs and create
the comparison.pdf}
\usage{
compScenEDGET(
  listofruns,
  hist,
  y_bar = c(2010, 2030, 2050, 2100),
  mainReg = "EUR",
  filename = "CompareScenarios_EDGET",
  load_cache = FALSE,
  cache_folder = "cache",
  AggrReg = "H12"
)
}
\arguments{
\item{listofruns}{folder hosting REMIND input files. If NULL, a list of magclass objects is returned (set this option in case of a REMIND preprocessing run)}

\item{hist}{path to a REMIND historical data file ("historical.mif")}

\item{y_bar}{years to plot on y axis}

\item{mainReg}{Main region to plot in single window}

\item{filename}{name of the pdf file to be produced}

\item{load_cache}{switch load from local cache on and off}

\item{cache_folder}{path to internal cache folder for local use}

\item{AggrReg}{Region aggregation for generating plot data}
}
\description{
Read in all information from previous EDGE runs and create
the comparison.pdf
}
\author{
Johanna Hoppe
}
