% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createRDS.R
\name{createRDS}
\alias{createRDS}
\title{Creates RDS files for iterative REMIND-EDGE runs from csv input files.
Existing files are overwritten silently. Does not return anything.}
\usage{
createRDS(input_path, data_path, SSP_scenario, EDGE_scenario)
}
\arguments{
\item{input_path}{the path to the folder containing the input (csv-) files}

\item{data_path}{the path to the output folder}

\item{SSP_scenario}{the SSP dimension, usually this refers to the GDP scenario in REMIND}

\item{EDGE_scenario}{the EDGE-T scenario dimension}
}
\description{
Creates RDS files for iterative REMIND-EDGE runs from csv input files.
Existing files are overwritten silently. Does not return anything.
}
