% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supportFunctionsIterative.R
\name{csv2RDS}
\alias{csv2RDS}
\title{Creates RDS files for iterative REMIND-EDGE runs from csv input files.
Existing files are overwritten silently. Does not return anything.}
\usage{
csv2RDS(filename, inputPath, SSPscenario, demScenario, transportPolScenario)
}
\arguments{
\item{filename}{name of the file}

\item{inputPath}{the path to the folder containing the input (csv-) files}

\item{SSPscenario}{SSP scenario}

\item{demScenario}{demand scenario}

\item{transportPolScenario}{Transport policy scenario}
}
\description{
Creates RDS files for iterative REMIND-EDGE runs from csv input files.
Existing files are overwritten silently. Does not return anything.
}
