% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edgeTransport-package.R
\docType{package}
\name{edgeTransport-package}
\alias{edgeTransport}
\alias{edgeTransport-package}
\title{edgeTransport: Prepare EDGE Transport Data for the REMIND model}
\description{
EDGE-T is a fork of the GCAM transport module https://jgcri.github.io/gcam-doc/energy.html#transportation with a high level of detail in its representation of technological and modal options. It is a partial equilibrium model with a nested multinomial logit structure and relies on the modified logit formulation. Most of the sources are not publicly available. PIK-internal users can find the sources in the distributed file system in the folder `/p/projects/rd3mod/inputdata/sources/EDGE-Transport-Standalone`.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/pik-piam/edgeTransport}
}

}
\author{
\strong{Maintainer}: Johanna Hoppe \email{johanna.hoppe@pik-potsdam.de} (\href{https://orcid.org/0009-0004-6753-5090}{ORCID})

Authors:
\itemize{
  \item Johanna Hoppe \email{johanna.hoppe@pik-potsdam.de} (\href{https://orcid.org/0009-0004-6753-5090}{ORCID})
  \item Alois Dirnaichner \email{dirnaichner@pik-potsdam.de}
  \item Marianna Rottoli \email{rottoli@pik-potsdam.de}
  \item Jarusch Muessel \email{jarusch.muessel@pik-potsdam.de} (\href{https://orcid.org/0000-0002-1857-7866}{ORCID})
  \item Alex K. Hagen \email{alex.hagen@pik-potsdam.de} (\href{https://orcid.org/0000-0003-4793-8664}{ORCID})
  \item Robert P. Pietzcker \email{pietzcker@pik-potsdam.de} (\href{https://orcid.org/0000-0002-9403-6711}{ORCID})
}

}
\keyword{internal}
