% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcEdgeTransportSA.R
\name{calcEdgeTransportSA}
\alias{calcEdgeTransportSA}
\title{Generate EDGE-Transport Input Data for the REMIND model, madrat interface.}
\usage{
calcEdgeTransportSA(
  SSPscen,
  transportPolScen,
  isICEban = FALSE,
  demScen = "default",
  gdxPath = NULL,
  outputFolder = NULL,
  isStored = FALSE,
  isTransportReported = TRUE,
  isTransportExtendedReported = FALSE,
  isREMINDinputReported = FALSE,
  isAnalyticsReported = FALSE
)
}
\arguments{
\item{SSPscen}{SSP or SDP scenario}

\item{transportPolScen}{EDGE-T transport policy scenario}

\item{isICEban}{optional enabling of ICE ban}

\item{demScen}{Demand scenario, used to apply reduction factors on total demands from the regression}

\item{gdxPath}{Path to a GDX file to load price signals from a REMIND run}

\item{outputFolder}{Path to folder for storing output data}

\item{isStored}{Optional saving of intermediate RDS files}

\item{isTransportReported}{Optional transport reporting in MIF format}

\item{isTransportExtendedReported}{Optional extension of transport reporting providing more detailed variables}

\item{isREMINDinputReported}{Optional reporting of REMIND input data}

\item{isAnalyticsReported}{Optional reporting of analytics data (e.g. variables over iterations)}
}
\value{
Transport input data for REMIND
}
\description{
\code{calcEdgeTransportSA()} is a wrapper for \code{toolEdgeTransportSA()} to make use of
madrat caching.
}
\author{
Jarusch Muessel, Johanna Hoppe
}
