% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateEDGEdata.R
\name{generateEDGEdata}
\alias{generateEDGEdata}
\title{Generate EDGE-Transport Input Data for the REMIND model.}
\usage{
generateEDGEdata(
  input_folder,
  output_folder,
  SSP_scen = "SSP2",
  tech_scen = "Mix",
  smartlifestyle = FALSE,
  storeRDS = FALSE,
  loadLvl0Cache = FALSE
)
}
\arguments{
\item{input_folder}{folder hosting raw data}

\item{output_folder}{folder hosting REMIND input files. If NULL, a list of magclass objects is returned (set this option in case of a REMIND preprocessing run)}

\item{SSP_scen}{SSP or SDP scenario}

\item{tech_scen}{EDGE-T technology scenario. Options are: ConvCase, ElecEra, HydrHype (working with SSP2 only!)}

\item{smartlifestyle}{If True, GDP demand regression provides lower overall demand levels.}

\item{storeRDS}{optional saving of intermediate RDS files, only possible if output folder is not NULL}

\item{loadLvl0Cache}{optional load intermediate RDS files for input data to save time}
}
\value{
generated EDGE-transport input data
}
\description{
Run this script to prepare the input data for EDGE in EDGE-friendly units and regional aggregation
}
\author{
Alois Dirnaichner, Marianna Rottoli
}
