% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supportFunctions.R
\name{getFilterEntriesUnivocalName}
\alias{getFilterEntriesUnivocalName}
\title{List associated univocalNames for any entry of the decision tree}
\usage{
getFilterEntriesUnivocalName(categories, decisionTree)
}
\arguments{
\item{categories}{vector of categories to filter}

\item{decisionTree}{decision tree that contains the univocalNames associated to the category}
}
\value{
list of categories and their associated univocalNames e.g. trn_pass as list entry
         containing a vector c("Compact Car", "HSR", ..)
}
\description{
List associated univocalNames for any entry of the decision tree
}
\author{
Johanna Hoppe
}
