% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createRDS.R
\name{loadInputData}
\alias{loadInputData}
\title{Load EDGE-T input data (RDS format) from a given path.}
\usage{
loadInputData(data_path)
}
\arguments{
\item{data_path}{path to RDS data files}
}
\value{
A list of data.tables
}
\description{
Load EDGE-T input data (RDS format) from a given path.
}
