% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lvl0_loadSources.R
\name{lvl0_CHNTrucksCosts}
\alias{lvl0_CHNTrucksCosts}
\title{Load the China truck costs}
\usage{
lvl0_CHNTrucksCosts(input_folder, years)
}
\arguments{
\item{input_folder}{folder hosting raw data}

\item{years}{time steps}
}
\value{
CAPEX and non-fuel OPEX conventional trucks
}
\description{
Final values:
  non fuel price in 2005USD/vkm
}
\author{
Marianna Rottoli
}
