% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lvl0_GCAM.R
\name{lvl0_correctTechOutput}
\alias{lvl0_correctTechOutput}
\title{Fix issues with the GCAM data}
\usage{
lvl0_correctTechOutput(tech_output)
}
\description{
Applies corrections to GCAM outdated/horrible data. No conversion of units happening.
}
