% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lvl0_GCAM.R
\name{lvl0_GCAMraw}
\alias{lvl0_GCAMraw}
\title{Read and prepare GCAM data}
\usage{
lvl0_GCAMraw(input_folder, GCAM2ISO_MAPPING, GDP_country, GCAM_dir = "GCAM")
}
\arguments{
\item{input_folder}{folder hosting raw data}

\item{GCAM2ISO_MAPPING}{a mapping between ISO3 codes and GCAM region names}

\item{GDP_country}{country level GDP PPP (used for disaggregation)}

\item{GCAM_dir}{subdirectory within the data input with GCAM data}
}
\value{
transport entries (demand, energy intensity, load factor, value of time, structure of the logit, vehicle speed)
}
\description{
Demand in million pkm and tmk, EI in MJ/km
}
