% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lvl0_IntAvPreparation.R
\name{lvl0_IntAvPreparation}
\alias{lvl0_IntAvPreparation}
\title{Loads and prepared historical international aviation demand data}
\usage{
lvl0_IntAvPreparation(input_folder, GDP_country, ICCT_dir = "ICCT")
}
\arguments{
\item{input_folder}{input folder}

\item{GDP_country}{GDP_country data}

\item{ICCT_dir}{Directory of the ICCT data}
}
\description{
Loads and prepared historical international aviation demand data
}
\author{
Sebastian Franz
}
