% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lvl0_TRACCS.R
\name{lvl0_loadEU}
\alias{lvl0_loadEU}
\title{Load EU data}
\usage{
lvl0_loadEU(input_folder, EU_dir = "TRACCS")
}
\arguments{
\item{input_folder}{folder hosting raw data}

\item{EU_dir}{directory with TRACCS data and Eurostat data}
}
\description{
Loads absolute values demand (pkm,tkm) for:
  Road (pass+freight), Rail (pass+freight), Shipping (freight), Aviation (pass)
Loads factors and Energy intensity for Road (pass+freight)
}
\details{
Final units for demand: millionkm (tkm and pkm)
}
