% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lvl0_TRACCS.R
\name{lvl0_loadTRACCS}
\alias{lvl0_loadTRACCS}
\title{Load TRACCS data}
\usage{
lvl0_loadTRACCS(input_folder, TRACCS_dir = "TRACCS")
}
\arguments{
\item{TRACCS_dir}{}
}
\description{
Loads absolute values demand (pkm,tkm) for:
  Road (pass+freight), Rail (pass+freight), Shipping (freight), Aviation (pass)
Loads factors and Energy intensity for Road (pass+freight)
}
\details{
Final units for demand: millionkm (tkm and pkm)
}
