% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lvl0_PSI.R
\name{lvl0_mergePSIintensity}
\alias{lvl0_mergePSIintensity}
\title{Merge PSI Energy Intensity for EU with GCAM Data}
\usage{
lvl0_mergePSIintensity(GCAM_data, input_folder, PSI_dir = "PSI")
}
\arguments{
\item{PSI_dir}{}
}
\description{
Final values: MJ/km (pkm and tkm)
}
