% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lvl0_TRACCS.R
\name{lvl0_mergeTRACCS}
\alias{lvl0_mergeTRACCS}
\title{Merge TRACCS data to ISO level output}
\usage{
lvl0_mergeTRACCS(TRACCS_data, output)
}
\description{
EI in MJ/km, LF in p/v, demand in million pkm or million tkm
}
