% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lvl0_mrremind.R
\name{lvl0_mrremind}
\alias{lvl0_mrremind}
\title{Read and prepare mrremind data}
\usage{
lvl0_mrremind(
  SSP_scen,
  REMIND2ISO_MAPPING,
  cache_folder,
  load_cache = FALSE,
  mrremind_folder = NULL
)
}
\arguments{
\item{SSP_scen}{SSP/SDP/other REMIND GDP scenario}

\item{REMIND2ISO_MAPPING}{mapping from REMIND regions to ISO3 country codes}

\item{cache_folder}{folder to use to store cache}

\item{load_cache}{(default=FALSE) load cached files from the input folder
to bypass the heavy mrremind machinery.}

\item{mrremind_folder}{folder hosting mrremind cache, required if load_cache=TRUE}
}
\value{
a list of mrremind-derived data, see end of this file.
}
\description{
Read and prepare mrremind data
}
\author{
Alois Dirnaichner, Marianna Rottoli
}
