% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lvl0_TRACCS.R
\name{lvl0_prepareEU}
\alias{lvl0_prepareEU}
\title{Prepare TRACCS and Eurostat data to merge them with other input data}
\usage{
lvl0_prepareEU(
  EU_data,
  iso_data,
  intensity,
  input_folder,
  GCAM2ISO_MAPPING,
  REMIND2ISO_MAPPING
)
}
\arguments{
\item{EU_data}{TRACCS based data}

\item{iso_data}{iso level data}

\item{intensity}{energy intensity of technologies}

\item{input_folder}{folder hosting raw data}

\item{GCAM2ISO_MAPPING}{GCAM2iso mapping}

\item{REMIND2ISO_MAPPING}{REMIND2iso mapping}
}
\description{
function that makes the TRACCS and Eurostat databases compatible with the rest of the data
Final values: EI in MJ/km (pkm and tkm), demand in million km (pkm and tkm), LF in p/v
}
