% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lvl0_TRACCS.R
\name{lvl0_prepareTRACCS}
\alias{lvl0_prepareTRACCS}
\title{Prepare TRACCS data}
\usage{
lvl0_prepareTRACCS(TRACCS_data, GCAM_data, intensity, input_folder,
  TRACCS_dir = "TRACCS")
}
\arguments{
\item{TRACCS_dir}{}
}
\description{
function that makes the TRACCS database compatible with the GCAM framework.
Final values: EI in MJ/km (pkm and tkm), demand in million km (pkm and tkm), LF in p/v
}
