% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lvl0_loadSources.R
\name{lvl0_PSIint}
\alias{lvl0_PSIint}
\title{Load the PSI energy intensity}
\usage{
lvl0_PSIint(GCAM_data, input_folder, PSI_dir = "PSI", years)
}
\arguments{
\item{GCAM_data}{GCAM data}

\item{input_folder}{folder hosting raw data}

\item{PSI_dir}{PSI data directory}

\item{years}{time steps}
}
\value{
CAPEX and non-fuel OPEX conventional trucks
}
\description{
Final values:
}
\author{
Marianna Rottoli, Alois Dirnaichner
}
