% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lvl1_calibrateEDGE.R
\name{lvl1_calibrateEDGE}
\alias{lvl1_calibrateEDGE}
\title{Calibrate the logit share weights to historical data.}
\usage{
lvl1_calibrateEDGE(prices, GCAM_data, logit_exp_data, vot_data)
}
\arguments{
\item{vot_data}{}
}
\description{
Calibrate the logit share weights to historical data.
}
