% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lvl1_calibrateEDGEinconv.R
\name{lvl1_calibrateEDGEinconv}
\alias{lvl1_calibrateEDGEinconv}
\title{Calibrate the logit share weights to historical data.}
\usage{
lvl1_calibrateEDGEinconv(
  prices,
  tech_output,
  logit_exp_data,
  vot_data,
  price_nonmot
)
}
\arguments{
\item{prices}{full prices (fuel, non-fuel) of each technology-vehicle type}

\item{tech_output}{regional historical demand}

\item{logit_exp_data}{logit exponents input data}

\item{vot_data}{value of time input data}

\item{price_nonmot}{non motorized technologies price}
}
\description{
Calibrate the logit share weights to historical data.
}
