% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lvl1_IEAharmonization.R
\name{lvl1_IEAharmonization}
\alias{lvl1_IEAharmonization}
\title{Harmonize the energy intensities to match the IEA energy balances.}
\usage{
lvl1_IEAharmonization(int, demKm, IEA)
}
\arguments{
\item{int}{intensity}

\item{demKm}{demand in million km}

\item{IEA}{IEA balances}
}
\description{
We provide energy service trajectories. IEA energy balances have to be met and are *not*
consistent with GCAM intensities and ES trajectories.
Therefore we have to change energy intensities for historical timesteps and smoothly
change them to the GCAM/PSI values.
}
