% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lvl1_SWtrend.R
\name{lvl1_SWclustering}
\alias{lvl1_SWclustering}
\title{Create regional clusters to bin different share weight trends.}
\usage{
lvl1_SWclustering(scenario)
}
\description{
Uses internal package data from WDI:
WDI_area_country=data.table(WDI::WDI(indicator = "AG.SRF.TOTL.K2",extra=TRUE))
}
