% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lvl2_finalSteps.R
\name{lvl2_createCSV}
\alias{lvl2_createCSV}
\title{Creates csv files that contains all EDGE-transport entries for Moinput.}
\usage{
lvl2_createCSV(logit_params, sw_data, vot_data, NEC_data, demByTech,
  int_dat, intensity, capCost, demand_traj, REMIND_scenario, EDGE_scenario,
  level2path)
}
\description{
Creates csv files that contains all EDGE-transport entries for Moinput.
}
