% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lvl2_finalSteps.R
\name{lvl2_createCSV_inconv}
\alias{lvl2_createCSV_inconv}
\title{Creates csv files that contains all EDGE-transport entries for mrremind.}
\usage{
lvl2_createCSV_inconv(
  logit_params,
  pref_data,
  vot_data,
  NEC_data,
  capcost4W,
  demByTech,
  int_dat,
  intensity,
  capCost,
  price_nonmot,
  complexValues,
  loadFactor,
  demISO,
  REMIND_scenario,
  EDGE_scenario,
  level2path
)
}
\arguments{
\item{logit_params}{logit parameters}

\item{pref_data}{preference factors}

\item{vot_data}{value of time}

\item{NEC_data}{non fuel cots}

\item{capcost4W}{capital costs 4wheelers}

\item{demByTech}{demand projections at REMIND level}

\item{int_dat}{intensity at technology level}

\item{intensity}{intensity at REMIND level}

\item{capCost}{capital costs}

\item{price_nonmot}{non motorized modes price}

\item{complexValues}{values for complex module in REMIND}

\item{loadFactor}{load factors}

\item{demISO}{ISO level demand in 2010 to be used as a weight in mrremind}

\item{REMIND_scenario}{SSP scenario}

\item{EDGE_scenario}{EDGE transport scenario specifier}

\item{level2path}{directory where data will be saved}
}
\description{
Creates csv files that contains all EDGE-transport entries for mrremind.
}
