% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lvl2_finalSteps.R
\name{lvl2_createoutput}
\alias{lvl2_createoutput}
\title{Creates csv files that contains all EDGE-transport entries for mrremind or a list of objects.}
\usage{
lvl2_createoutput(
  logit_params,
  pref_data,
  vot_data,
  NEC_data,
  capcost4W,
  demByTech,
  int_dat,
  intensity,
  capCost,
  price_nonmot,
  complexValues,
  loadFactor,
  annual_mileage,
  demISO,
  SSP_scen,
  EDGE_scenario,
  level2path,
  output_folder
)
}
\arguments{
\item{logit_params}{logit parameters}

\item{pref_data}{preference factors}

\item{vot_data}{value of time}

\item{NEC_data}{non fuel cots}

\item{capcost4W}{capital costs 4wheelers}

\item{demByTech}{demand projections at REMIND level}

\item{int_dat}{intensity at technology level}

\item{intensity}{intensity at REMIND level}

\item{capCost}{capital costs}

\item{price_nonmot}{non motorized modes price}

\item{complexValues}{values for complex module in REMIND}

\item{loadFactor}{load factors}

\item{annual_mileage}{annual mileage}

\item{demISO}{ISO level demand in 2010 to be used as a weight in mrremind}

\item{SSP_scen}{SSP scenario}

\item{EDGE_scenario}{EDGE transport scenario specifier}

\item{level2path}{directory where data will be saved}

\item{output_folder}{directory where the data has to be saved, if set to NULL a list of magclass objects will be saved}
}
\description{
Creates csv files that contains all EDGE-transport entries for mrremind or a list of objects.
}
