% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lvl2_demandReg.R
\name{lvl2_demandReg}
\alias{lvl2_demandReg}
\title{Estimate future transport demand using a price trajectory, GDP, and historical demand.}
\usage{
lvl2_demandReg(tech_output, price_baseline, REMIND_scenario)
}
\arguments{
\item{price_baseline}{}
}
\description{
Estimate future transport demand using a price trajectory, GDP, and historical demand.
}
