% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lvl2_demandRegNAVIGATEIntl.R
\name{lvl2_demandRegNAVIGATEIntl}
\alias{lvl2_demandRegNAVIGATEIntl}
\title{Estimate future transport demand using a price trajectory, GDP, and historical demand.}
\usage{
lvl2_demandRegNAVIGATEIntl(
  tech_output,
  price_baseline,
  GDP_POP,
  ICCT_data,
  RPK_cap_baseline,
  input_folder,
  COVID_dir = "COVID",
  REMIND_scenario,
  smartlifestyle,
  Baseline_Run
)
}
\arguments{
\item{tech_output}{historically calibrated demand}

\item{price_baseline}{baseline prices}

\item{GDP_POP}{GDP per capita}

\item{ICCT_data}{data on international aviation}

\item{RPK_cap_baseline}{baseline demand}

\item{input_folder}{folder where the input data is}

\item{COVID_dir}{folder where input data is about COVID}

\item{REMIND_scenario}{SSP scenario}

\item{smartlifestyle}{switch activating sustainable lifestyles}

\item{Baseline_Run}{switch: is it a baseline or an intl aviation run}
}
\value{
transport demand projections
}
\description{
Estimate future transport demand using a price trajectory, GDP, and historical demand.
}
\author{
Sebastian Franz
}
