% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lvl2_generatePlotdata.R
\name{lvl2_generate_plotdata}
\alias{lvl2_generate_plotdata}
\title{Generate EDGE-Transport Plot Data for Standalone Report}
\usage{
lvl2_generate_plotdata(listofruns, AggrReg = "H12")
}
\arguments{
\item{listofruns}{list of folders hosting raw data}

\item{AggrReg}{Regional Aggregation}
}
\value{
generated EDGE-transport plot data
}
\description{
Run this script to prepare the input data for EDGE in EDGE-friendly units and regional aggregation
}
\author{
Johanna Hoppe
}
