% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_prices.R
\name{merge_prices}
\alias{merge_prices}
\title{Merge REMIND-derived fuel prices with non-fuel costs.}
\usage{
merge_prices(
  gdx,
  REMINDmapping,
  REMINDyears,
  intensity_data,
  nonfuel_costs,
  module = "edge_esm"
)
}
\arguments{
\item{gdx}{path to REMIND binary output, gdx format}

\item{REMINDmapping}{mapping of REMIND regions to ISO3 country codes}

\item{REMINDyears}{range of REMIND timesteps}

\item{intensity_data}{logit level intensity data}

\item{nonfuel_costs}{logti level non-fuel costs}

\item{module}{GDX input is based on old "complex" module or new "edge_esm" module}
}
\description{
Merge REMIND-derived fuel prices with non-fuel costs.
}
