% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportEDGETransport2.R
\name{reportEDGETransport2}
\alias{reportEDGETransport2}
\title{Reporting for the coupled EDGE-T Transport Sector Model (REMIND Module edge_esm)}
\usage{
reportEDGETransport2(
  output_folder = ".",
  extendedReporting = FALSE,
  scenario_title = NULL,
  model_name = "EDGE-Transport",
  gdx = NULL
)
}
\arguments{
\item{output_folder}{path to the output folder, default is current folder.}

\item{extendedReporting}{report a larger set of variables}

\item{scenario_title}{a scenario title string}

\item{model_name}{a model name string}

\item{gdx}{path to the GDX file used for the run.}
}
\description{
Data is loaded from the EDGE-T subfolder in the output folder or from the `level_2` folder
in a standalone run.
For a coupled run the EDGE-T output files can be (re-) generated calling
`Rscript EDGETransport.R --reporting`
from the output folder.
}
\details{
Region subsets are obtained from fulldata.gdx
}
\author{
Johanna Hoppe Alois Dirnaichner Marianna Rottoli
}
