% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolAdjustDemandCovid.R
\name{toolAdjustDemandCovid}
\alias{toolAdjustDemandCovid}
\title{Function to adjust the sector ES demand on CES level to the covid decrease in 2020}
\usage{
toolAdjustDemandCovid(demandData)
}
\arguments{
\item{demandData}{the energy service demand on CES level: trn_aviation_intl, trn_shipping_intl, trn_freight, trn_pass}
}
\value{
demandData adjusted for the decrease in demand related to covid
}
\description{
The adjustment is based on the relative differences of our FE model projections
in 2020 to the FE IEA data averaged over five years.
In addition, the deviation in 2020 is corrected for earlier diferences between IEA data and model projections,
such that only the decrease in deviation due to covid is regarded here. 
The final differences used for this correction are  stored in inst/extdata/IEAdemandDifferencesCovid.csv
For more context, the full data and calculations can be found 
in /p/projects/edget/adjustmentDataFiles/IEAdemandDifferencesCovid_compHistoricREMIND.xlsx
}
\details{
The demand projections are constantly shifted by an absolute difference from 2020 onwards

This function is applied in toolDemandRegression() directly after the regression calculation as a default
}
