% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjustIntensity.R
\name{toolAdjustIntensity}
\alias{toolAdjustIntensity}
\title{Adjust energy intensity on technology level based on techSen specific annual improvement rates and given start and end year}
\usage{
toolAdjustIntensity(intensity, intImproTab, years)
}
\arguments{
\item{intensity}{energy intensity raw data in EJ/pkm or EJ/tkm}

\item{intImproTab}{table that provides techScen specific annual improvement rates, start year and end year}

\item{years}{time horizon of EDGE-T}
}
\value{
Adjusted energy intensities in EJ/pkm or EJ/tkm
}
\description{
Adjust energy intensity on technology level based on techSen specific annual improvement rates and given start and end year
}
