% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolApplyICEbanOnPreferences.R
\name{toolApplyICEbanOnPreferences}
\alias{toolApplyICEbanOnPreferences}
\title{Apply ICE ban on vehicle types that feature preference factors}
\usage{
toolApplyICEbanOnPreferences(preferenceTab, helpers, ICEbanYears)
}
\arguments{
\item{preferenceTab}{data.table including preferences for all levels of the decision tree}

\item{helpers}{list of helpers}

\item{ICEbanYears}{sequence of years in which ICEban is applied}
}
\value{
Preferences in accordance to the ICE ban policy
}
\description{
Apply ICE ban on vehicle types that feature preference factors
}
\author{
Johanna Hoppe, Alex K. Hagen
}
