% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supportFunctions.R
\name{toolApplyMixedTimeRes}
\alias{toolApplyMixedTimeRes}
\title{toolApplyMixedTimeRes}
\usage{
toolApplyMixedTimeRes(data, helpers, idcols = NULL)
}
\arguments{
\item{data}{data.table containing data in in the edgeTransport data structure
(at least featuring univocalName, period, value)}

\item{helpers}{list containing several helpers used throughout the model.
It includes dtTimeRes, a data.table containing the temporal
resolution for different univocalNames}

\item{idcols}{optional supply of idcols for using approx_dt}
}
\value{
data.table
}
\description{
Applies two different temporal resolutions on a data.table object in the
edgeTransport data structure and do a linear approximination for the highRes data that is not available
}
\author{
Johanna Hoppe
}
