% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolApplyScenPrefTrends.R
\name{toolApplyScenPrefTrends}
\alias{toolApplyScenPrefTrends}
\title{Apply scenario specific adjustments to the preference trends}
\usage{
toolApplyScenPrefTrends(
  baselinePrefTrends,
  scenParPrefTrends,
  GDPpcMER,
  policyStartYear,
  GDPcutoff,
  helpers,
  isICEban
)
}
\arguments{
\item{baselinePrefTrends}{Baseline preference trends}

\item{scenParPrefTrends}{Scenario parameters to be applied on the preference trends}

\item{GDPpcMER}{Per capita GDP based on market exchange rate}

\item{policyStartYear}{Year from which the scenario parameters are applied on the baseline preference trends}

\item{GDPcutoff}{Threshold used to categorize countries into different mitigation groups based on their GDP}

\item{helpers}{List containing several helpers used throughout the model}

\item{isICEban}{Switch to turn on ICE phase out policies}
}
\value{
Scenario specific preference trends
}
\description{
Apply scenario specific adjustments to the preference trends
}
\author{
Johanna Hoppe
}
