% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolApplyScenSpecEnInt.R
\name{toolApplyScenSpecEnInt}
\alias{toolApplyScenSpecEnInt}
\title{Apply demand scenario specific adjustments to the energy intensity}
\usage{
toolApplyScenSpecEnInt(enInt, scenParEnergyIntensity, policyStartYear, helpers)
}
\arguments{
\item{enInt}{Energy intensity input data supplied by mrtransport}

\item{scenParEnergyIntensity}{Transport policy scenario specific energy intensity improvement factors}

\item{policyStartYear}{Year from which scenario specific transport policies are applied}

\item{helpers}{List with helpers}
}
\value{
data.table with scenario specific energy intensity input data
}
\description{
Apply demand scenario specific adjustments to the energy intensity
}
\author{
Johanna Hoppe
}
