% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolApplyScenSpecLoadFactor.R
\name{toolApplyScenSpecLoadFactor}
\alias{toolApplyScenSpecLoadFactor}
\title{Apply demand scenario specific adjustments to the load Factor}
\usage{
toolApplyScenSpecLoadFactor(
  loadFactor,
  scenParLoadFactor,
  policyStartYear,
  helpers
)
}
\arguments{
\item{loadFactor}{load factor input data supplied by mrtransport}

\item{scenParLoadFactor}{Scenario specific parameters to be applied don the baseline load factor}

\item{policyStartYear}{Year when scenario differentiation sets in}

\item{helpers}{List with helpers}
}
\value{
Scenario specific load factor
}
\description{
Apply demand scenario specific adjustments to the load Factor
}
\author{
Johanna Hoppe
}
