% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolCalculateAnnuity.R
\name{toolCalculateAnnuity}
\alias{toolCalculateAnnuity}
\title{Calculate annuity for different vehicle types}
\usage{
toolCalculateAnnuity(annuityCalc, helpers)
}
\arguments{
\item{annuityCalc}{input data for interest rate and service Life from edgeTransport}

\item{mitigationTechMap}{map to disaggregate vehicle types}
}
\description{
Calculate annuity for different vehicle types
}
\author{
Johanna Hoppe
}
