% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolCalculateFleetComposition.R
\name{toolCalculateFleetComposition}
\alias{toolCalculateFleetComposition}
\title{Calculate composition of the fleet}
\usage{
toolCalculateFleetComposition(
  ESdemandFVsalesLevel,
  vehDepreciationFactors,
  vehSalesAndModeShares,
  annualMileage,
  loadFactor,
  helpers
)
}
\arguments{
\item{ESdemandFVsalesLevel}{Energy service demand on fuel vehicle level for new sales}

\item{vehDepreciationFactors}{Factors for the depreciation of vehicles in each year of the maximum service Life}

\item{vehSalesAndModeShares}{Vehicle sales and mode shares for the decision tree}

\item{annualMileage}{Annual kilometers driven by a vehicle}

\item{loadFactor}{Persons or tons loaded on each vehicle}

\item{helpers}{List of helpers}
}
\value{
list of fleet composition data
}
\description{
This function calculates the fleet composition based on the energy service demand, the composition of new sales
each year and the service life of vehicles. With the help of the annual mileage, the absolute numbers of vehicles is calculated.
If the remaining stock is sufficient to meet the energy service demand, 10% of the fleet are retired early to ensure a minimum sales volume each year.
}
\author{
Johanna Hoppe
}
