% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolCalculateInitialIncoCost.R
\name{toolCalculateInitialIncoCost}
\alias{toolCalculateInitialIncoCost}
\title{Apply regional differences for inconvenience cost start values based on ICE cost differences}
\usage{
toolCalculateInitialIncoCost(
  combinedCost,
  incoCostStartVal,
  annuity,
  loadFactor,
  annualMileage,
  helpers
)
}
\arguments{
\item{combinedCost}{total cost of ownership}

\item{incoCostStartVal}{start values for inconvenience costs}

\item{annuity}{calculated annuity for different vehicle types}

\item{loadFactor}{load factor data}

\item{annualMileage}{annual mileage data}

\item{helpers}{list with helpers}
}
\value{
data.table including initial inconvenience costs from 1990-2020 for LDV 4W US$2017/(p|t)km
}
\description{
Apply regional differences for inconvenience cost start values based on ICE cost differences
}
\author{
Johanna Hoppe
}
