% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logit.R
\name{toolCalculateLogitIncost}
\alias{toolCalculateLogitIncost}
\title{Evaluate 4W logit tree shares from real prices and inconvenience costs.
Inconvenience costs are updated based on shares from previous timesteps (endogenous formulation).}
\usage{
toolCalculateLogitIncost(
  prices,
  vot_data,
  pref_data,
  logit_params,
  intensity_data,
  price_nonmot,
  ptab4W,
  tech_scen,
  totveh = NULL
)
}
\arguments{
\item{prices}{logit prices}

\item{vot_data}{value-of-time data}

\item{pref_data}{inconvenience cost data and SWs}

\item{logit_params}{contains logit exponents}

\item{intensity_data}{logit level intensity data}

\item{price_nonmot}{price of non-motorized modes in the logit tree}

\item{ptab4W}{inconvenience cost factors for LDVs}

\item{tech_scen}{technology scenario}

\item{totveh}{total demand for LDVs by tecnology, in million veh}
}
\description{
Evaluate 4W logit tree shares from real prices and inconvenience costs.
Inconvenience costs are updated based on shares from previous timesteps (endogenous formulation).
}
