% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logit.R
\name{toolCalculateLogitSW}
\alias{toolCalculateLogitSW}
\title{Evaluate logit tree shares from real prices and inconvenience costs.
Inconvenience costs are updated based on shares from previous timesteps (endogenous formulation).}
\usage{
toolCalculateLogitSW(
  prices,
  vot_data,
  pref_data,
  logit_params,
  intensity_data,
  price_nonmot,
  ptab4W,
  logit_data_4W
)
}
\arguments{
\item{prices}{logit prices}

\item{vot_data}{value-of-time data}

\item{pref_data}{inconvenience cost data and SWs}

\item{logit_params}{contains logit exponents}

\item{intensity_data}{logit level intensity data}

\item{price_nonmot}{price of non-motorized modes in the logit tree}

\item{ptab4W}{inconvenience cost factors for LDVs}

\item{logit_data_4W}{the output of the 4W logit calculation}
}
\description{
Evaluate logit tree shares from real prices and inconvenience costs.
Inconvenience costs are updated based on shares from previous timesteps (endogenous formulation).
}
