% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolCalculateSharesDecisionTree.R
\name{toolCalculateSharesDecisionTree}
\alias{toolCalculateSharesDecisionTree}
\title{Calculate shares for all levels of the decision tree for a variable on FV level}
\usage{
toolCalculateSharesDecisionTree(dtVariable, helpers)
}
\arguments{
\item{dtVariable}{transport variable e.g. energy service demand on the most granular level fuel-vehicle (FV)}

\item{helpers}{description}
}
\value{
data.table including shares of the variable for all levels of the decision tree
}
\description{
Calculate shares for all levels of the decision tree for a variable on FV level
}
\author{
Johanna Hoppe
}
