% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolCalculateVehicleDepreciationFactors.R
\name{toolCalculateVehicleDepreciationFactors}
\alias{toolCalculateVehicleDepreciationFactors}
\title{Calculate vehicle depreciation factors for fleet tracking}
\usage{
toolCalculateVehicleDepreciationFactors(annuityCalc, helpers)
}
\arguments{
\item{annuityCalc}{parameters for annualization}

\item{helpers}{list with helpers}
}
\value{
data.table including vehicle depreciation factors over service life#'
}
\description{
Calculate vehicle depreciation factors for fleet tracking
}
\author{
Johanna Hoppe
}
