% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolCalibrateHistPrefs.R
\name{toolCalibrateHistPrefs}
\alias{toolCalibrateHistPrefs}
\title{Calibrate the logit share weights to historical data.}
\usage{
toolCalibrateHistPrefs(
  combinedCosts,
  histESdemand,
  timeValueCost,
  lambdas,
  helpers
)
}
\arguments{
\item{combinedCosts}{Annualized total cost of ownership}

\item{histESdemand}{Historical energy service demand data}

\item{timeValueCost}{Time value cost for passenger transport modes}

\item{lambdas}{Exponents for discrete choice function}

\item{helpers}{list with helpers}
}
\value{
data.table with calibrated historical preferences
}
\description{
Calibrate the logit share weights to historical data.
}
