% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolCalibratePreferences.R
\name{toolCalibratePreferences}
\alias{toolCalibratePreferences}
\title{Calibrate the logit share weights to historical data.}
\usage{
toolCalibratePreferences(
  sharesToBeCalibrated,
  combinedCosts,
  timeValueCost,
  lambdas,
  helpers
)
}
\arguments{
\item{sharesToBeCalibrated}{Shares at each level of the decision tree for which preferences are to be calibrated}

\item{combinedCosts}{Annualized total cost of ownership}

\item{timeValueCost}{Time value cost for passenger transport modes}

\item{lambdas}{Exponents for discrete choice function}

\item{helpers}{list with helpers}
}
\value{
data.table with calibrated historical preferences
}
\description{
Calibrate the logit share weights to historical data.
}
