% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supportFunctions.R
\name{toolCheckAllLevelsComplete}
\alias{toolCheckAllLevelsComplete}
\title{toolCheckAllLevelsComplete}
\usage{
toolCheckAllLevelsComplete(data, decisionTree, name)
}
\arguments{
\item{data}{data.table containing data in all levels format that should be checked}

\item{decisionTree}{data.table containing full edgeTransport decision Tree}

\item{name}{name of variable to be checked}
}
\value{
data.table
}
\description{
Checks whether data is complete for all levels of decision tree
}
\author{
Johanna Hoppe
}
