% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadSources.R
\name{toolCHNTrucksCosts}
\alias{toolCHNTrucksCosts}
\title{Load the China truck costs}
\usage{
toolCHNTrucksCosts(input_folder, years)
}
\arguments{
\item{input_folder}{folder hosting raw data}

\item{years}{time steps}
}
\value{
CAPEX and non-fuel OPEX conventional trucks
}
\description{
Final values:
  non fuel price in 2005USD/vkm
}
\author{
Marianna Rottoli
}
