% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolCombineCAPEXandOPEX.R
\name{toolCombineCAPEXandOPEX}
\alias{toolCombineCAPEXandOPEX}
\title{Function that converts CAPEX and OPEX into US$2005/(p|t)km and provides them combined in a structured format}
\usage{
toolCombineCAPEXandOPEX(
  CAPEXtrackedFleet,
  nonFuelOPEXtrackedFleet,
  CAPEXother,
  nonFuelOPEXother,
  fuelCosts,
  subsidies,
  energyIntensity,
  loadFactor,
  annualMileage,
  annuity,
  helpers
)
}
\arguments{
\item{fuelCosts}{fuel costs from REMIND}

\item{subsidies}{subsides for alternative cars from mrremind}

\item{annuity}{calculated annuity for different vehicle types}

\item{mrtransportInput}{list containing all mrtransport input data}

\item{decisionTree}{edgeTransport decision tree}

\item{filter}{list of filters for specific branches in the upper decision tree, containing all associated univocalNames}
}
\value{
data.table including total costs of ownership in US$2005/(p|t)km
}
\description{
Function that converts CAPEX and OPEX into US$2005/(p|t)km and provides them combined in a structured format
}
