% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolCombineCAPEXandOPEXiterative.R
\name{toolCombineCAPEXandOPEXiterative}
\alias{toolCombineCAPEXandOPEXiterative}
\title{Function that converts CAPEX and OPEX into US$/(p|t)km and provides them combined in a structured format}
\usage{
toolCombineCAPEXandOPEXiterative(
  CAPEXandNonFuelOPEX,
  fuelCosts,
  energyIntensity,
  loadFactor,
  annualMileage,
  helpers
)
}
\arguments{
\item{CAPEXandNonFuelOPEX}{non-fuel OPEX and CAPEX data as combined in the previous EDGET run and loaded from RDS file}

\item{fuelCosts}{updated fuel cost from REMIND}

\item{energyIntensity}{energy intensity data}

\item{loadFactor}{load factor data}

\item{annualMileage}{annual mileage data}

\item{helpers}{list with helpers}
}
\value{
data.table including total costs of ownership in US$/(p|t)km
}
\description{
Function that converts CAPEX and OPEX into US$/(p|t)km and provides them combined in a structured format
}
