% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supportFunctionsIterative.R
\name{toolCreateRDS}
\alias{toolCreateRDS}
\title{Creates RDS files for iterative REMIND-EDGE runs from csv input files.
Existing files are overwritten silently. Does not return anything.}
\usage{
toolCreateRDS(
  inputPath,
  edgeTransportFolder,
  SSPscen,
  demScen,
  transportPolScen
)
}
\arguments{
\item{inputPath}{the path to the folder containing the input (csv-) files}

\item{demScen}{the SSP dimension, usually this refers to the GDP scenario in REMIND}

\item{transportPolScen}{the EDGE-T scenario dimension}

\item{dataPath}{the path to the output folder}

\item{DEM_scenario}{the demand scenario.}
}
\description{
Creates RDS files for iterative REMIND-EDGE runs from csv input files.
Existing files are overwritten silently. Does not return anything.
}
