% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demandRegression.R
\name{toolDemandReg}
\alias{toolDemandReg}
\title{Estimate future transport demand using a price trajectory, GDP, and historical demand.}
\usage{
toolDemandReg(
  tech_output,
  price_baseline,
  GDP_POP,
  SSP_scen,
  ssp_factors,
  regional_factors = NULL,
  demscen_factors = NULL
)
}
\arguments{
\item{tech_output}{historically calibrated demand}

\item{price_baseline}{baseline prices}

\item{GDP_POP}{GDP per capita}

\item{SSP_scen}{REMIND SSP scenario}

\item{ssp_factors}{table with SSP/SDP scenario factors to be added to the baseline elasticity trends}

\item{regional_factors}{table with SSP/SDP and regionally specific factors to be added to the baseline elasticity trends}

\item{demscen_factors}{table with reduction factors on the total demands,
example: demscen_factors <-
fread("demandScen,  region,sector,  year, factor
       SSP2EU_lowEn,DEU,   trn_pass,2030, 0.5")}
}
\value{
transport demand projections
}
\description{
Estimate future transport demand using a price trajectory, GDP, and historical demand.
}
\author{
Marianna Rottoli, Alois Dirnaichner
}
