% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolDiscreteChoice.R
\name{toolDiscreteChoice}
\alias{toolDiscreteChoice}
\title{Calculate vehicle sales shares and mode shares for all levels of the decisionTree.}
\usage{
toolDiscreteChoice(input, generalModelPar, updatedEndoCosts, helpers)
}
\arguments{
\item{input}{dataset for discrete choice module}

\item{generalModelPar}{general model parameter}

\item{updatedEndoCosts}{updated endogenous costs}

\item{helpers}{list of helpers}
}
\value{
calculated shares
}
\description{
Function that traverses each level of the decision tree and calculates the shares according to the discrete choice approach
}
\author{
Johanna Hoppe
}
