% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateEDGEdata.R
\name{toolGenerateEDGEdata}
\alias{toolGenerateEDGEdata}
\title{Generate EDGE-Transport Input Data for the REMIND model.}
\usage{
toolGenerateEDGEdata(
  input_folder,
  output_folder,
  cache_folder = NULL,
  SSP_scen = "SSP2",
  tech_scen = "Mix1",
  demScen = "default",
  storeRDS = FALSE,
  gdxPath = NULL,
  preftab = NULL,
  plot.report = FALSE,
  mitab4W.path = NULL,
  mitab.path = NULL,
  ssp_demreg.path = NULL,
  regional_demreg.path = NULL,
  demscen.path = NULL,
  FEPricetab = NULL,
  int_improvetab = NULL
)
}
\arguments{
\item{input_folder}{folder hosting raw data}

\item{output_folder}{folder hosting REMIND input files. If NULL, a list of magclass objects is returned (set this option in case of a REMIND preprocessing run)}

\item{cache_folder}{folder hosting a "local" cache (this is not the mrremid cache, it is specific to EDGE-T).}

\item{SSP_scen}{SSP or SDP scenario}

\item{tech_scen}{EDGE-T technology scenario. Options are: ConvCase, ElecEra, HydrHype (working with SSP2 only!)}

\item{demScen}{Demand scenario, used to apply reduction factors on total demands from the regression.}

\item{storeRDS}{optional saving of intermediate RDS files, only possible if output folder is not NULL}

\item{gdxPath}{optional path to a GDX file to load price signals from a REMIND run.}

\item{preftab}{path to file with trends for share weights}

\item{plot.report}{write a report which is place in the level2 folder. Defaults to FALSE.}

\item{mitab4W.path}{path to file with key factors for 4W technologies for different mitigation ambition and SSP scenarios.}

\item{mitab.path}{path to file with key factors for share weight trends for different mitigation ambition and SSP scenarios.}

\item{ssp_demreg.path}{path to file with key factors for the demand regression depending on SSP/SDP scenarios.}

\item{regional_demreg.path}{path to file with key factors for the demand regression depending on regions and SSP scenarios.}

\item{demscen.path}{path to file with demScen specific mulitiplikative factors for the demand, region and sector specific.}

\item{FEPricetab}{ship an external csv that includes FE prices. The prices from the gdx file will be overwritten for affected regions.}

\item{int_improvetab}{table with key factors for intensity improvements for different mitigation ambition and SSP scenarios.}
}
\value{
generated EDGE-transport input data
}
\description{
Run this script to prepare the input data for EDGE in EDGE-friendly units and regional aggregation
}
\details{
\code{calcgenerateEDGEdata()} is a wrapper for \code{generateEDGEdata()} to make use of
madrat caching.
}
\author{
Alois Dirnaichner, Marianna Rottoli
}
