% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolGetCommonParameters.R
\name{toolGetCommonParameters}
\alias{toolGetCommonParameters}
\title{Distribute common parameters to standalone and iterative EDGET script}
\usage{
toolGetCommonParameters(startyear, ICEbanBefore = FALSE, ICEbanAfter = FALSE)
}
\arguments{
\item{startyear}{First time point in which policy differentiation sets in, cm_startyear in REMIND}

\item{ICEbanBefore}{Boolean to determine if ICEban is in effect before startyear}

\item{ICEbanAfter}{Boolean to determine if ICEban is in effect from startyear on}
}
\value{
a list of parameters: allEqYear, ICEbanYears, GDPcutoff, baseYear, hybridElecShare
}
\description{
Distribute common parameters to standalone and iterative EDGET script
}
\author{
Alex K. Hagen
}
