% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/incocost.R
\name{toolIncocost}
\alias{toolIncocost}
\title{Calculates the inconvenience costs for 4wheelers}
\usage{
toolIncocost(annual_mileage, load_factor, fcr_veh, REMINDp)
}
\arguments{
\item{annual_mileage}{annual mileage of vehicles}

\item{load_factor}{load factor vehicles}

\item{fcr_veh}{depreciation rate}

\item{REMINDp}{REMIND prices}
}
\description{
Calculates the inconvenience costs for 4wheelers
}
