% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supportFunctions.R
\name{toolLoadDecisionTree}
\alias{toolLoadDecisionTree}
\title{Read and build the complete structure of the edgeTransport decision tree}
\usage{
toolLoadDecisionTree(regionAggregation = "iso")
}
\arguments{
\item{regionAggregation}{choose one of the different options for regional aggregation (iso|regionCode21|regionCode12)}
}
\value{
data.table of full spatially extended edgeTransport decision tree
}
\description{
Read and build the complete structure of the edgeTransport decision tree
}
\author{
Johanna Hoppe
}
