% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supportFunctionsIterative.R
\name{toolLoadIterativeInputs}
\alias{toolLoadIterativeInputs}
\title{Load iterative inputs}
\usage{
toolLoadIterativeInputs(
  edgeTransportFolder,
  inputFolder,
  inputFiles,
  numberOfRegions,
  SSPscenario,
  transportPolScenario,
  demScenario
)
}
\arguments{
\item{edgeTransportFolder}{transport folder}

\item{inputFolder}{the path to the folder containing the input (csv-) files}

\item{inputFiles}{names of the input files}

\item{numberOfRegions}{regional resolution}

\item{SSPscenario}{SSP scenario}

\item{transportPolScenario}{Transport policy scenario}

\item{demScenario}{demand scenario}
}
\description{
Load iterative inputs
}
