% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolLoadmrdriversData.R
\name{toolLoadmrdriversData}
\alias{toolLoadmrdriversData}
\title{Load GDP and Population data from mrdrivers}
\usage{
toolLoadmrdriversData(SSPscen, helpers)
}
\arguments{
\item{SSPscen}{SSP scenario for which the mrdrivers input data shall be loaded}

\item{helpers}{list containg several helpers used throughout the model.
It includes dtTimeRes, a data.table containing the temporal
resolution for different univocalNames}
}
\value{
list of data.tables containing mrdrivers input data
}
\description{
Load GDP and Population data from mrdrivers
}
