% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolLoadmrremindData.R
\name{toolLoadmrremindData}
\alias{toolLoadmrremindData}
\title{Load data from mrremind}
\usage{
toolLoadmrremindData(helpers)
}
\arguments{
\item{helpers}{list containg several helpers used throughout the model.
It includes dtTimeRes, a data.table containing the temporal
resolution for different univocalNames}
}
\value{
list of data.tables containing mrremind input data
}
\description{
Load data from mrremind
}
