% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolLoadmrtransportData.R
\name{toolLoadmrtransportData}
\alias{toolLoadmrtransportData}
\title{Load input data from the mrtransport package}
\usage{
toolLoadmrtransportData(SSPscen)
}
\arguments{
\item{SSPscen}{SSP scenario for which the mrtransport input data should be loaded}
}
\value{
list of data.tables with the mrtransport input data
}
\description{
Load input data from the mrtransport package
}
