% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolLoadRDSinputs.R
\name{toolLoadRDSinputs}
\alias{toolLoadRDSinputs}
\title{Creates RDS files for iterative REMIND-EDGE runs from csv input files.
Existing files are overwritten silently. Does not return anything.}
\usage{
toolLoadRDSinputs(edgeTransportFolder, inputFiles)
}
\arguments{
\item{edgeTransportFolder}{directory in the output folder of the REMIND run that stores EDGET RDS files}

\item{inputFiles}{list of filename strings that should be read in}
}
\description{
Creates RDS files for iterative REMIND-EDGE runs from csv input files.
Existing files are overwritten silently. Does not return anything.
}
