% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolLoadREMINDesDemand.R
\name{toolLoadREMINDesDemand}
\alias{toolLoadREMINDesDemand}
\title{Load REMIND energy service demand}
\usage{
toolLoadREMINDesDemand(gdxPath, helpers)
}
\arguments{
\item{gdxPath}{path to REMIND fulldata.gdx}

\item{helpers}{list of helpers}
}
\description{
Load the energy service demand from a REMIND fulldata.gdx on sector level (CES level)
in [bn (p|t)km/yr] and the requested temporal resolution
}
