% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolLoadScenParIncoCost.R
\name{toolLoadScenParIncoCost}
\alias{toolLoadScenParIncoCost}
\title{Function to load a parameter set for a specific scenario(combination) from the csv file in the package}
\usage{
toolLoadScenParIncoCost(SSPs, transportPolS)
}
\arguments{
\item{SSPs}{SSP scenarios}

\item{transportPolS}{transport policy scenarios}
}
\value{
list with different input data sets
}
\description{
Function to load a parameter set for a specific scenario(combination) from the csv file in the package
}
\author{
Alex K. Hagen
}
