% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadSources.R
\name{toolLoadUCD}
\alias{toolLoadUCD}
\title{Load the UCD transportation database}
\usage{
toolLoadUCD(input_folder, fcr_veh, years, UCD_dir = "UCD")
}
\arguments{
\item{input_folder}{folder hosting raw data}

\item{fcr_veh}{annualization factor for LDVs}

\item{years}{time steps}

\item{UCD_dir}{directory with UCD data}
}
\value{
non fuel price and annual mileage of vehicles
}
\description{
Final values:
  non fuel price in 2005$/vkm
  annual mileage in vkt/veh/yr (vehicle km traveled per year),
}
\author{
Marianna Rottoli
}
