% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrremind.R
\name{toolMrremind}
\alias{toolMrremind}
\title{Read and prepare mrremind data}
\usage{
toolMrremind(SSP_scen, REMIND2ISO_MAPPING, cache_folder)
}
\arguments{
\item{SSP_scen}{SSP/SDP/other REMIND GDP scenario}

\item{REMIND2ISO_MAPPING}{mapping from REMIND regions to ISO3 country codes}

\item{cache_folder}{folder to use to store and load inputdata cache}
}
\value{
a list of mrremind-derived data, see end of this file.
}
\description{
Read and prepare mrremind data
}
\author{
Alois Dirnaichner, Marianna Rottoli
}
