% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supportFunctions.R
\name{toolNormalizePreferences}
\alias{toolNormalizePreferences}
\title{Normalize preferences so that the maximum in each branch of the decision tree equals 1}
\usage{
toolNormalizePreferences(preferenceTab)
}
\arguments{
\item{preferenceTab}{data.table including preferences for all levels of the decision tree}
}
\value{
Normalized preferences
}
\description{
Normalize preferences so that the maximum in each branch of the decision tree equals 1
}
\author{
Johanna Hoppe
}
