% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supportFunctions.R
\name{toolOrderandCheck}
\alias{toolOrderandCheck}
\title{toolOrderandCheck}
\usage{
toolOrderandCheck(
  data,
  decisionTree,
  yrs = NULL,
  checkCompleteness = FALSE,
  fleetVars = FALSE
)
}
\arguments{
\item{data}{data.table containing data in all levels format that should be checked}

\item{decisionTree}{data.table containing full edgeTransport decision Tree}

\item{checkCompleteness}{if activated it is further checked, if the full data set is provided}

\item{fleetVars}{if activated the full data set is reduced to vehicle types that feature fleet tracking}
}
\value{
data.table
}
\description{
sort data.table according to edgeTransport data structure and check for NAs
If checkCompletness is activated it is further checked, if the full data set is provided.
In case of fleetVars == TRUE the full data set is reduced to the vehicle types that feature fleet tracking
}
\author{
Johanna Hoppe
}
