% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/incotrend.R
\name{toolPreftrend}
\alias{toolPreftrend}
\title{Calculate a trend for share weights and inconvenience costs based on the EDGE scenario}
\usage{
toolPreftrend(
  SWS,
  ptab,
  calibdem,
  incocost,
  years,
  GDP_POP_MER,
  tech_scen,
  SSP_scen,
  mitab
)
}
\arguments{
\item{SWS}{preference factors}

\item{ptab}{mode and veh preferences table}

\item{calibdem}{calibration demand}

\item{incocost}{inconvenience costs for 4wheelers}

\item{years}{time steps}

\item{GDP_POP_MER}{GDP population on MER base}

\item{tech_scen}{technology at the center of the policy packages}

\item{SSP_scen}{SSP or SDP scenario}

\item{mitab}{mitigation pathways table}
}
\value{
projected trend of preference factors
}
\description{
Calculate a trend for share weights and inconvenience costs based on the EDGE scenario
}
\author{
Alois Dirnaichner, Marianna Rottoli
}
