% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolPrepareDataEndogenousCosts.R
\name{toolPrepareDataEndogenousCosts}
\alias{toolPrepareDataEndogenousCosts}
\title{Format all cost components}
\usage{
toolPrepareDataEndogenousCosts(inputData, lambdas, policyStartYear, helpers)
}
\arguments{
\item{inputData}{List containing inputData}

\item{lambdas}{exponents for discrete choice calculation}

\item{policyStartYear}{Year when scenario differentiation sets in}

\item{helpers}{List with helpers}
}
\value{
data.table including all cost components
}
\description{
Format all cost components
}
\author{
Johanna Hoppe
}
