% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadSources.R
\name{toolPSICosts}
\alias{toolPSICosts}
\title{Load the PSI costs}
\usage{
toolPSICosts(input_folder, years, fcr_veh)
}
\arguments{
\item{input_folder}{folder hosting raw data}

\item{years}{time steps}

\item{fcr_veh}{annualization factor for LDVs}
}
\value{
purchase cost LDVs
}
\description{
Final values:
  non fuel price in 2005USD annualized
}
\author{
Marianna Rottoli
}
