% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readREMINDdemand.R
\name{toolReadREMINDdemand}
\alias{toolReadREMINDdemand}
\title{Loads the total demand of transport CES nodes from last iteration}
\usage{
toolReadREMINDdemand(gdx, REMINDmapping, EDGE2teESmap, years, scenario)
}
\arguments{
\item{gdx}{REMIND binary output, gdx format}

\item{REMINDmapping}{map REMIND regions to ISO3 country codes}

\item{EDGE2teESmap}{map EDGE-T technologies to REMIND ES techs}

\item{years}{REMIND timesteps}

\item{scenario}{REMIND SSP scenario}
}
\description{
Loads the total demand of transport CES nodes from last iteration
}
