% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolReLoadInputs.R
\name{toolReLoadInputs}
\alias{toolReLoadInputs}
\title{called from iterativeEdgeTransport()
reloads scenario specific inputData from RDS files in EDGE-T folder
instead of generation via mrtransport}
\usage{
toolReLoadInputs(edgeTransportFolder)
}
\arguments{
\item{edgeTransportFolder}{folder where the RDS files from last iterativeEdgeTransport() run are stored}
}
\value{
list with different input data sets
}
\description{
called from iterativeEdgeTransport()
reloads scenario specific inputData from RDS files in EDGE-T folder
instead of generation via mrtransport
}
\author{
Alex K. Hagen
}
