% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finalSteps.R
\name{toolREMINDdemand}
\alias{toolREMINDdemand}
\title{Provide demand trajectories for REMIND}
\usage{
toolREMINDdemand(regrdemand, EDGE2teESmap, REMINDtall, SSP_scen)
}
\arguments{
\item{regrdemand}{demand regression}

\item{EDGE2teESmap}{mapping between EDGE-T and REMIND technologies}

\item{REMINDtall}{extended REMIND time steps}

\item{SSP_scen}{SSP scenario}
}
\description{
Provide demand trajectories for REMIND
}
