% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolTranslateTransportScenario.R
\name{toolTranslateTransportScenario}
\alias{toolTranslateTransportScenario}
\title{Function to rename scenarios between REMIND and EDGE-Transport naming conventions
It is used at the interfaces REMIND-EDGETiterative, EDGETiterative-REMIND
The same translations have to be done in mrremind readEDGETransport for REMIND input data generation}
\usage{
toolTranslateTransportScenario(
  demScen,
  transportPolScen,
  direction = c("EDGEtoREMIND", "REMINDtoEDGE")
)
}
\arguments{
\item{demScen}{demand scenario,
if demScen is in mapping it will be translated otherwise returned unchanged}

\item{transportPolScen}{transport policy scenario
if transportPolScen is in mapping it will be translated otherwise returned unchanged}

\item{direction}{one of two labels indicating the direction of the translation}
}
\value{
list of translated demScen and transportPolScen in the desired model name logic
}
\description{
Function to rename scenarios between REMIND and EDGE-Transport naming conventions
It is used at the interfaces REMIND-EDGETiterative, EDGETiterative-REMIND
The same translations have to be done in mrremind readEDGETransport for REMIND input data generation
}
\author{
Alex K. Hagen
}
