% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supportFunctions.R
\name{toolTraverseDecisionTree}
\alias{toolTraverseDecisionTree}
\title{Calculate data for the next higher level of the decision tree.}
\usage{
toolTraverseDecisionTree(data, upperLevel, decisionTree)
}
\arguments{
\item{data}{data.table containing the data of the lower level}

\item{upperLevel}{name of the upper level in the decision tree}

\item{decisionTree}{full edgeTransport decision tree}
}
\value{
data.table
}
\description{
Function that aggregates the data for the next higher level of the decision tree
}
\author{
Johanna Hoppe
}
