% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolUpdateEndogenousCosts.R
\name{toolUpdateEndogenousCosts}
\alias{toolUpdateEndogenousCosts}
\title{toolUpdateEndogenousCosts}
\usage{
toolUpdateEndogenousCosts(
  dataEndoCosts,
  depreciationFactors,
  scenParIncoCost,
  policyStartYear,
  timeValue,
  preferences,
  lambdas,
  helpers,
  isICEban,
  vehiclesPerTech = NULL
)
}
\arguments{
\item{dataEndoCosts}{data.table containing all cost components for cars over the full range
of policy years in a yearly resolution. Exogenous CAPEX and OPEX are provided over the full range.
Endogenous cost components and FS3 shares are provided until policyStartYear. Rest is filled with NA.}

\item{depreciationFactors}{data.table containing vehicle depreciation factor for each year of service Life}

\item{scenParIncoCost}{data.table containing scenario specific parameters for inconvenience costs policy mask}

\item{policyStartYear}{year from which scenario-specific differentiation begins}

\item{timeValue}{data.table containing mode specific time value costs based on speed and gdp}

\item{preferences}{preference factor trends}

\item{lambdas}{data.table containing exponents for discrete choice calculation}

\item{helpers}{list containing helpers like mappings, decisionTree etc.}

\item{isICEban}{optional enabling of ICE ban}

\item{vehiclesPerTech}{data.table containing total number of vehicles for all years and regions}
}
\value{
list containing data.table with endogenous cost components over the full time span and additional data.tables
        for model behavior analysis
}
\description{
Provides updates for endogenous cost components e.g. inconvenience costs for cars
}
\author{
Johanna Hoppe
}
