% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gcam.R
\name{toolVOTandExponents}
\alias{toolVOTandExponents}
\title{Load value-of-time and exponents}
\usage{
toolVOTandExponents(
  GCAM_data,
  GDP_MER_country,
  POP_country,
  input_folder,
  logitexp_dir = "GCAM_logit_exponents"
)
}
\arguments{
\item{GCAM_data}{GCAM based data}

\item{GDP_MER_country}{GDP iso level MER}

\item{POP_country}{population (ISO level)}

\item{input_folder}{folder hosting raw data}

\item{logitexp_dir}{directory with logit exponents for GCAM}
}
\description{
load logit exponents for each level: they are based on GCAM assumptions.
They are on csv files that already follow the EDGE structure, created by hand.
}
\details{
VOT values in (1990$/pkm)
}
