% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateEDGEdata.R
\name{generateEDGEdata}
\alias{generateEDGEdata}
\title{Generate EDGE-Transport Input Data for the REMIND model.}
\usage{
generateEDGEdata(
  input_folder,
  output_folder,
  EDGE_scenario,
  REMIND_scenario = "SSP2",
  IEAbal = NULL,
  GDP_country = NULL,
  POP_country = NULL,
  JRC_IDEES_Trsp = NULL,
  JRC_IDEES_MarBunk = NULL,
  saveRDS = FALSE
)
}
\arguments{
\item{input_folder}{folder hosting raw data}

\item{output_folder}{folder hosting REMIND input files}

\item{EDGE_scenario}{EDGE transport scenario specifier}

\item{REMIND_scenario}{SSP scenario}

\item{IEAbal}{use mrremind generated data: in case of a REMIND preprocessing run, load population.  Product of: calcOutput("IO", subtype = "IEA_output", aggregate = TRUE)}

\item{GDP_country}{use mrremind generated data: in case of a REMIND preprocessing run, load GDP.  Product of: calcOutput("GDPppp", aggregate =F)}

\item{POP_country}{use mrremind generated data: in case of a REMIND preprocessing run, load IEA balances. Product of: calcOutput("Population", aggregate =F)}

\item{JRC_IDEES_Trsp}{use mrremind generated data: in case of a REMIND preprocessing run, load JRC_IDEE data. Product of: calcOutput("JRC_IDEES", subtype="Transport")}

\item{JRC_IDEES_MarBunk}{use mmrremind generated data: in case of a REMIND preprocessing run, load JRC_IDEE data. Product of: calcOutput("JRC_IDEES", subtype="maritimeBunkers")}

\item{saveRDS}{optional saving of intermediate RDS files}
}
\value{
generated EDGE-transport input data
}
\description{
Run this script to prepare the input data for EDGE in EDGE-friendly units and regional aggregation
}
\author{
Alois Dirnaichner, Marianna Rottoli
}
