% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateEDGEdata.R
\name{generateEDGEdata}
\alias{generateEDGEdata}
\title{Generate EDGE-Transport Input Data for the REMIND model.}
\usage{
generateEDGEdata(
  input_folder,
  output_folder,
  cache_folder = "cache",
  SSP_scen = "SSP2",
  tech_scen = "Mix",
  smartlifestyle = FALSE,
  val_excel_input = FALSE,
  storeRDS = FALSE,
  loadLvl0Cache = FALSE,
  gdxPath = NULL,
  preftab = NULL,
  plot.report = FALSE,
  mitab4W.path = NULL,
  mitab.path = NULL
)
}
\arguments{
\item{input_folder}{folder hosting raw data}

\item{output_folder}{folder hosting REMIND input files. If NULL, a list of magclass objects is returned (set this option in case of a REMIND preprocessing run)}

\item{cache_folder}{folder hosting a "local" cache (this is not the mrremid cache, it is specific to EDGE-T). NOTE: the cache folder will be created in the output_folder if it does not exist.}

\item{SSP_scen}{SSP or SDP scenario}

\item{tech_scen}{EDGE-T technology scenario. Options are: ConvCase, ElecEra, HydrHype (working with SSP2 only!)}

\item{smartlifestyle}{If True, GDP demand regression provides lower overall demand levels.}

\item{val_excel_input}{switch: if TRUE, the validation routine of preferences takes place.}

\item{storeRDS}{optional saving of intermediate RDS files, only possible if output folder is not NULL}

\item{loadLvl0Cache}{optional load intermediate RDS files for input data to save time}

\item{gdxPath}{optional path to a GDX file to load price signals from a REMIND run.}

\item{preftab}{path to file with trends for share weights}

\item{plot.report}{write a report which is place in the level2 folder. Defaults to FALSE.}

\item{mitab4W.path}{path to file with key factors for 4W technologies for different mitigation ambition and SSP scenarios.}

\item{mitab.path}{path to file with key factors for share weight trends for different mitigation ambition and SSP scenarios.}
}
\value{
generated EDGE-transport input data
}
\description{
Run this script to prepare the input data for EDGE in EDGE-friendly units and regional aggregation
}
\author{
Alois Dirnaichner, Marianna Rottoli
}
