% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateEDGEdata.R
\name{generateEDGEdata}
\alias{generateEDGEdata}
\title{Generate EDGE-Transport Input Data for the REMIND model.}
\usage{
generateEDGEdata(
  input_folder,
  output_folder,
  SSP_scen = "SSP2",
  techscen = NULL,
  IEAbal = NULL,
  GDP_country = NULL,
  RatioPPP2MER_country = NULL,
  POP_country = NULL,
  JRC_IDEES_Trsp = NULL,
  JRC_IDEES_MarBunk = NULL,
  trsp_incent = NULL,
  storeRDS = FALSE,
  loadRDS = FALSE
)
}
\arguments{
\item{input_folder}{folder hosting raw data}

\item{output_folder}{folder hosting REMIND input files. If NULL, a list of magclass objects is returned (set this option in case of a REMIND preprocessing run)}

\item{SSP_scen}{SSP or SDP scenario}

\item{techscen}{EDGE-T technology scenario. Options are: ConvCase, ElecEra, HydrHype (working with SSP2 only!)}

\item{IEAbal}{use mrremind generated data: in case of a REMIND preprocessing run, load population.  Product of: calcOutput("IO", subtype = "IEA_output", aggregate = TRUE)}

\item{GDP_country}{use mrremind generated data: in case of a REMIND preprocessing run, load GDP.  Product of: calcOutput("GDPppp", aggregate =F)}

\item{RatioPPP2MER_country}{use mrremind generated data: in case of a REMIND preprocessing run, load ratio between PPP and MER GDP.  Product of: calcOutput("RatioPPP2MER", aggregate =F)}

\item{POP_country}{use mrremind generated data: in case of a REMIND preprocessing run, load IEA balances. Product of: calcOutput("Population", aggregate =F)}

\item{JRC_IDEES_Trsp}{use mrremind generated data: in case of a REMIND preprocessing run, load JRC_IDEE data. Product of: calcOutput("JRC_IDEES", subtype="Transport")}

\item{JRC_IDEES_MarBunk}{use mmrremind generated data: in case of a REMIND preprocessing run, load JRC_IDEE data. Product of: calcOutput("JRC_IDEES", subtype="maritimeBunkers")}

\item{trsp_incent}{use mmrremind generated data: in case of a REMIND preprocessing run, load transportSubsidies data. Product of: readSource("TransportSubsidies", convert=T)}

\item{storeRDS}{optional saving of intermediate RDS files, only possible if output folder is not NULL}

\item{loadRDS}{optional load intermediate RDS files for input data to save time}
}
\value{
generated EDGE-transport input data
}
\description{
Run this script to prepare the input data for EDGE in EDGE-friendly units and regional aggregation
}
\author{
Alois Dirnaichner, Marianna Rottoli
}
