% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateEDGEdata.R
\name{generateEDGEdata}
\alias{generateEDGEdata}
\title{Generate EDGE-Transport Input Data for the REMIND model.}
\usage{
generateEDGEdata(
  input_folder,
  output_folder,
  EDGE_scenario,
  REMIND_scenario = "SSP2",
  saveRDS = FALSE
)
}
\arguments{
\item{input_folder}{folder hosting raw data}

\item{output_folder}{folder hosting REMIND input files}

\item{EDGE_scenario}{EDGE transport scenario specifier}

\item{REMIND_scenario}{SSP scenario}

\item{saveRDS}{optional saving of intermediate RDS files}
}
\value{
generated EDGE-transport input data
}
\description{
Run this script to prepare the input data for EDGE in EDGE-friendly units and regional aggregation
}
\author{
Alois Dirnaichner, Marianna Rottoli
}
