% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iterativeEDGETransport.R
\name{iterativeEdgeTransport}
\alias{iterativeEdgeTransport}
\title{EDGE-Transport iterative
EDGE-Transport iterative script}
\usage{
iterativeEdgeTransport()
}
\description{
Run in the REMIND output folder in between iterations
}
\author{
Johanna Hoppe
}
