% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lvl0_GCAM.R
\name{lvl0_correctTechOutput}
\alias{lvl0_correctTechOutput}
\title{Fix issues with the GCAM data}
\usage{
lvl0_correctTechOutput(GCAM_output, NEcost, logitexp)
}
\arguments{
\item{GCAM_output}{output from GCAM raw data wrangling}

\item{NEcost}{non energy costs}

\item{logitexp}{logit exponents}
}
\description{
Applies corrections to GCAM outdated data. No conversion of units happening.
}
