% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lvl0_inconvCalc.R
\name{lvl0_incocost}
\alias{lvl0_incocost}
\title{Calculates the inconvenience costs for 4wheelers}
\usage{
lvl0_incocost(annual_mileage, load_factor, fcr_veh)
}
\arguments{
\item{annual_mileage}{annual mileage of vehicles}

\item{load_factor}{load factor vehicles}

\item{fcr_veh}{depreciation rate}
}
\description{
Calculates the inconvenience costs for 4wheelers
}
