% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lvl0_UCD.R
\name{lvl0_loadUCD}
\alias{lvl0_loadUCD}
\title{Load the UCD transportation database}
\usage{
lvl0_loadUCD(GCAM_data, EDGE_scenario, REMIND_scenario, input_folder,
  years, UCD_dir = "UCD")
}
\arguments{
\item{UCD_dir}{}
}
\description{
Loads and prepares the non_fuel prices.
It also load PSI-based purchase prices for EU.
Final values:
  non fuel price in 1990$/pkt (1990USD/tkm),
  annual mileage in vkt/veh/yr (vehicle km traveled per year),
  non_fuel_split in 1990$/pkt (1990USD/tkm)
}
