% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lvl0_toISO.R
\name{lvl0_toISO}
\alias{lvl0_toISO}
\title{Produce ISO versions of all files.}
\usage{
lvl0_toISO(
  input_data,
  VOT_data,
  price_nonmot,
  UCD_data,
  GDP,
  GDP_country,
  POP,
  GDP_POP,
  GCAM2ISO_MAPPING,
  REMIND2ISO_MAPPING,
  EDGE_scenario,
  REMIND_scenario = "SSP2"
)
}
\arguments{
\item{input_data}{demand and energy intensity input data}

\item{VOT_data}{value of time data}

\item{price_nonmot}{price non motorized transport}

\item{UCD_data}{UCD data}

\item{GDP}{regional level}

\item{GDP_country}{GDP ISO level}

\item{POP}{population regional level}

\item{GDP_POP}{per capita GDP}

\item{GCAM2ISO_MAPPING}{GCAM2iso mapping}

\item{REMIND2ISO_MAPPING}{regional mapping}

\item{EDGE_scenario}{EDGE transport scenario specifier}

\item{REMIND_scenario}{SSP scenario}
}
\value{
ISO level data for all entries
}
\description{
No conversion of units happening.
}
\author{
Marianna Rottoli
}
