% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lvl1_incotrend.R
\name{lvl1_preftrend}
\alias{lvl1_preftrend}
\title{Calculate a trend for share weights and inconvenience costs based on the EDGE scenario}
\usage{
lvl1_preftrend(
  SWS,
  calibdem,
  incocost,
  years,
  GDP,
  GDP_POP_MER,
  smartlifestyle,
  tech_scen,
  SSP_scen
)
}
\arguments{
\item{SWS}{preference factors}

\item{calibdem}{calibration demand}

\item{incocost}{inconvenience costs for 4wheelers}

\item{years}{time steps}

\item{GDP}{GDP regional level}

\item{GDP_POP_MER}{GDP per capita MER}

\item{smartlifestyle}{switch activating sustainable lifestyles}

\item{tech_scen}{technology at the center of the policy packages}

\item{SSP_scen}{REMIND SSP scenario}
}
\value{
projected trend of preference factors
}
\description{
Calculate a trend for share weights and inconvenience costs based on the EDGE scenario
}
\author{
Alois Dirnaichner, Marianna Rottoli
}
