% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lvl1_SWtrend.R
\name{lvl1_SWclustering}
\alias{lvl1_SWclustering}
\title{Create regional clusters to bin different share weight trends.}
\usage{
lvl1_SWclustering(
  input_folder,
  REMIND_scenario,
  REMIND2ISO_MAPPING,
  WDI_dir = "WDI"
)
}
\arguments{
\item{input_folder}{folder hosting raw data}

\item{REMIND_scenario}{SSP scenario}

\item{REMIND2ISO_MAPPING}{REMIND2iso mapping}

\item{WDI_dir}{directory with WDI data}
}
\value{
clusters grouping regions for preference factors trends
}
\description{
Create regional clusters to bin different share weight trends.
}
\author{
Marianna Rottoli
}
