% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lvl2_demandReg.R
\name{lvl2_demandReg}
\alias{lvl2_demandReg}
\title{Estimate future transport demand using a price trajectory, GDP, and historical demand.}
\usage{
lvl2_demandReg(tech_output, price_baseline, GDP_POP, smartlifestyle, SSP_scen)
}
\arguments{
\item{tech_output}{historically calibrated demand}

\item{price_baseline}{baseline prices}

\item{GDP_POP}{GDP per capita}

\item{smartlifestyle}{switch activating sustainable lifestyles}

\item{SSP_scen}{REMIND SSP scenario}
}
\value{
transport demand projections
}
\description{
Estimate future transport demand using a price trajectory, GDP, and historical demand.
}
\author{
Marianna Rottoli
}
