% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lvl2_finalSteps.R
\name{lvl2_REMINDdemand}
\alias{lvl2_REMINDdemand}
\title{Provide demand trajectories for REMIND}
\usage{
lvl2_REMINDdemand(regrdemand, EDGE2teESmap, REMINDtall, REMIND_scenario)
}
\arguments{
\item{regrdemand}{demand regression}

\item{EDGE2teESmap}{mapping between EDGE-T and REMIND technologies}

\item{REMINDtall}{extended REMIND time steps}

\item{REMIND_scenario}{SSP scenario}
}
\description{
Provide demand trajectories for REMIND
}
