% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lvl2_reportingEntries.R
\name{lvl2_reportingEntries}
\alias{lvl2_reportingEntries}
\title{Provides entries that are useful for reporting and complex realization}
\usage{
lvl2_reportingEntries(ESdem, FEdem)
}
\arguments{
\item{ESdem}{energy services demand}

\item{FEdem}{final energy demand}
}
\description{
Provides entries that are useful for reporting and complex realization
}
