% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolCalculateFS3share.R
\name{toolCalculateFS3share}
\alias{toolCalculateFS3share}
\title{toolCalculateFS3share}
\usage{
toolCalculateFS3share(
  endoCostData,
  timesteps,
  timeValue,
  preferences,
  lambdas,
  helpers
)
}
\arguments{
\item{endoCostData}{data.table containing all cost components on technology level}

\item{timesteps}{years for which to calculate FS3 shares}

\item{timeValue}{data.table containing mode specific time value costs based on speed and gdp}

\item{lambdas}{data.table containing exponents for discrete choice calculation}

\item{helpers}{list containing helpers like mappings, decisionTree etc.}
}
\value{
data.table containing all cost components on technology level and their respective FS3 shares
}
\description{
Provides updates for endogenous cost components e.g. inconvenience costs for cars
}
\author{
Johanna Hoppe
}
