% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supportFunctions.R
\name{toolCalculateSharesDecisionTree}
\alias{toolCalculateSharesDecisionTree}
\title{Calculate shares for all levels of the decision tree for a variable on FV level}
\usage{
toolCalculateSharesDecisionTree(dtVariable, helpers)
}
\arguments{
\item{dtVariable}{data.table of a transport variable, e.g., energy service demand with data at the most detailed level of of the decision tree (FV = fuel-vehicle)}

\item{helpers}{list of helpers}
}
\value{
data.table including shares of the variable for all levels of the decision tree
}
\description{
Calculate shares for all levels of the decision tree for a variable on FV level
}
\author{
Johanna Hoppe
}
