% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolCombineCAPEXandOPEX.R
\name{toolCombineCAPEXandOPEX}
\alias{toolCombineCAPEXandOPEX}
\title{Function that converts CAPEX and OPEX into US$2017/(p|t)km and provides them combined in a structured format}
\usage{
toolCombineCAPEXandOPEX(
  CAPEXtrackedFleet,
  nonFuelOPEXtrackedFleet,
  CAPEXother,
  nonFuelOPEXother,
  fuelCosts,
  subsidies,
  energyIntensity,
  loadFactor,
  annualMileage,
  annuity,
  helpers
)
}
\arguments{
\item{CAPEXtrackedFleet}{CAPEX data for vehicle types that feature fleet tracking: Cars, trucks, busses}

\item{nonFuelOPEXtrackedFleet}{non-fuel OPEX data for vehicle types that feature fleet tracking: Cars, trucks, busses}

\item{CAPEXother}{CAPEX data for other vehicle types}

\item{nonFuelOPEXother}{non-fuel OPEX data for other vehicle types}

\item{fuelCosts}{fuel cost data}

\item{subsidies}{purchase price subsidy data}

\item{energyIntensity}{energy intensity data}

\item{loadFactor}{load factor data}

\item{annualMileage}{annual mileage data}

\item{annuity}{calculated annuity for different vehicle types}

\item{helpers}{list with helpers}
}
\value{
data.table including total costs of ownership in US$2017/(p|t)km
}
\description{
Function that converts CAPEX and OPEX into US$2017/(p|t)km and provides them combined in a structured format
}
