% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supportFunctionsIterative.R
\name{toolCreateRDS}
\alias{toolCreateRDS}
\title{Creates RDS files for iterative REMIND-EDGE runs from csv input files.
Existing files are overwritten silently. Does not return anything.}
\usage{
toolCreateRDS(
  inputPath,
  edgeTransportFolder,
  inputFiles,
  SSPscen,
  demScen,
  transportPolScen
)
}
\arguments{
\item{inputPath}{the path to the folder containing the input (csv-) files}

\item{edgeTransportFolder}{the path to the EDGE-Transport folde}

\item{inputFiles}{files to load as input for the iterative script}

\item{SSPscen}{SSP scenario}

\item{demScen}{demand scenario}

\item{transportPolScen}{Transport policy scenario}
}
\description{
Creates RDS files for iterative REMIND-EDGE runs from csv input files.
Existing files are overwritten silently. Does not return anything.
}
