% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolDemandRegression.R
\name{toolDemandRegression}
\alias{toolDemandRegression}
\title{Energy service demand projection}
\usage{
toolDemandRegression(
  historicalESdemand,
  GDPperCapitaPPP,
  POP,
  scenParDemandRegression,
  scenParRegionalDemRegression,
  scenParDemandFactors,
  baseYear,
  policyStartYear,
  helpers
)
}
\arguments{
\item{historicalESdemand}{Historical energy service demand}

\item{GDPperCapitaPPP}{GDP per capita based on purchase power parity}

\item{POP}{Population data}

\item{scenParDemandRegression}{Scenario specific general regression factors}

\item{scenParRegionalDemRegression}{Scenario specific regionally differentiated regression factors}

\item{scenParDemandFactors}{Demand scenario factors}

\item{baseYear}{End year of historical energy service demand data}

\item{policyStartYear}{Start year of scenario differentiation}

\item{helpers}{list with helpers}
}
\value{
Scenario specific energy service demand for all model years on CES level
}
\description{
Energy service demand projection
}
\author{
Johanna Hoppe
}
