% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolEdgeTransportSA.R
\name{toolEdgeTransportSA}
\alias{toolEdgeTransportSA}
\title{Energy Demand Generator (EDGE)- Transport Model}
\usage{
toolEdgeTransportSA(
  SSPscen,
  transportPolScen,
  isICEban = c(FALSE, FALSE),
  demScen = c("default", "default"),
  policyStartYear = 2025,
  gdxPath = NULL,
  outputFolder = NULL,
  isStored = TRUE,
  isTransportReported = TRUE,
  isTransportExtendedReported = FALSE,
  isREMINDinputReported = FALSE,
  isAnalyticsReported = FALSE
)
}
\arguments{
\item{SSPscen}{SSP or SDP scenarios}

\item{transportPolScen}{EDGE-T transport policy scenarios}

\item{isICEban}{optional enabling of ICE ban}

\item{demScen}{Demand scenarios, used to apply reduction factors on total demands from the regression}

\item{policyStartYear}{Year after which policy differentiation sets in}

\item{gdxPath}{Path to a GDX file to load price signals from a REMIND run}

\item{outputFolder}{Path to folder for storing output data}

\item{isStored}{Optional saving of intermediate RDS files}

\item{isTransportReported}{Optional transport reporting in MIF format}

\item{isTransportExtendedReported}{Optional extension of transport reporting providing more detailed variables}

\item{isREMINDinputReported}{Optional reporting of REMIND input data}

\item{isAnalyticsReported}{Optional reporting of analytics data (e.g. variables over iterations)}
}
\value{
Transport input data for REMIND
}
\description{
The Edge Transport Model includes the transport specific input data preparation,
a choice model to determine transport mode and technology shares, a demand regression
and a fleet tracking for cars, busses and trucks
}
\author{
Johanna Hoppe, Jarusch Müßel, Alois Dirnaichner, Marianna Rottoli
}
