% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iterativeEDGETransport.R
\name{toolIterativeEDGETransport}
\alias{toolIterativeEDGETransport}
\title{EDGE-Transport iterative}
\usage{
toolIterativeEDGETransport()
}
\arguments{
\item{reporting}{generate EDGE-T reporting data}
}
\description{
Run in the REMIND output folder.
}
\author{
Alois Dirnaichner
}
