% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolLoadInputs.R
\name{toolLoadInputs}
\alias{toolLoadInputs}
\title{Load all inputs that are required to run the model}
\usage{
toolLoadInputs(SSPscen, transportPolScen, demScen, hybridElecShare, allEqYear)
}
\arguments{
\item{SSPscen}{SSP or SDP scenarios}

\item{transportPolScen}{EDGE-T transport policy scenarios}

\item{demScen}{Demand scenario, used to apply reduction factors on total demands from the regression}

\item{hybridElecShare}{Share of electricity in Hybrid electric vehicles}

\item{allEqYear}{Year after which scenario differentiation sets in, last year in which scenarios are the same}
}
\value{
list with different input data sets
}
\description{
Load all inputs that are required to run the model
}
\author{
Johanna Hoppe, Alex K. Hagen
}
