% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolLoadInputs.R
\name{toolLoadInputs}
\alias{toolLoadInputs}
\title{Load all inputs that are required to run the model}
\usage{
toolLoadInputs(SSPscen, transportPolScen, demScen, hybridElecShare)
}
\arguments{
\item{SSPscen}{SSP or SDP scenario}

\item{transportPolScen}{EDGE-T transport policy scenario}

\item{demScen}{Demand scenario, used to apply reduction factors on total demands from the regression}

\item{hybridElecShare}{Share of electricity in Hybrid electric vehicles}
}
\value{
list with different input data sets
}
\description{
Load all inputs that are required to run the model
}
\author{
Johanna Hoppe
}
