% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolLoadPackageData.R
\name{toolLoadPackageData}
\alias{toolLoadPackageData}
\title{Load input data from the edgeTransport package and choose data
according to SSP and transport policy scenario}
\usage{
toolLoadPackageData(SSPs, transportPolS, demScenario = NULL)
}
\arguments{
\item{SSPs}{SSP scenarios for which the package data should be loaded}

\item{transportPolS}{transport policy scenarios for which the package data should be loaded}

\item{demScenario}{demand scenario for which the package input data should be loaded}
}
\value{
list of data.tables with the package input data
}
\description{
Load input data from the edgeTransport package and choose data
according to SSP and transport policy scenario
}
