% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolReLoadInputs.R
\name{toolReLoadInputs}
\alias{toolReLoadInputs}
\title{When running with REMIND, starting from the second run of
iterativeEdgeTransport() scenario specific inputData can be reloaded from
saved files}
\usage{
toolReLoadInputs(edgeTransportFolder)
}
\arguments{
\item{edgeTransportFolder}{folder where the RDS files from last iterativeEdgeTransport() run are stored}
}
\value{
list with different input data sets
}
\description{
When running with REMIND, starting from the second run of
iterativeEdgeTransport() scenario specific inputData can be reloaded from
saved files
}
\author{
Alex K. Hagen
}
