% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolTranslateDemScen.R
\name{toolTranslateDemScen}
\alias{toolTranslateDemScen}
\title{Function to rename demand scenarios between REMIND and EDGE-Transport naming conventions
It is used at the interfaces REMIND-EDGETiterative, EDGETiterative-REMIND}
\usage{
toolTranslateDemScen(
  demScen,
  transportPolScen = NULL,
  isICEban = NULL,
  direction = c("EDGEtoREMIND", "REMINDtoEDGE")
)
}
\arguments{
\item{demScen}{demand scenario, to be renamed}

\item{transportPolScen}{transport policy scenario as it is used in EDGET}

\item{isICEban}{boolean indicationg if ICEban is on or off}

\item{direction}{one of two labels indicating the direction of the translation}
}
\value{
demScen in the desired model name logic
}
\description{
Function to rename demand scenarios between REMIND and EDGE-Transport naming conventions
It is used at the interfaces REMIND-EDGETiterative, EDGETiterative-REMIND
}
\author{
Alex K. Hagen
}
